feat(AccountSelector): Gestion des notifications initiales + réorganisation du code dans `AccountSelector`
